+++ src/compiler/llvm_codegen_value.c
@@ -42,22 +42,26 @@ void llvm_value_set(BEValue *value, LLVMValueRef llvm_value, Type *type)
 	}
 }
 
-void llvm_value_set_address(BEValue *value, LLVMValueRef llvm_value, Type *type, AlignSize alignment)
+void llvm_value_set_address(GenContext *c, BEValue *value, LLVMValueRef llvm_value, Type *type, AlignSize alignment)
 {
 	if (alignment == 0)
 	{
 		puts("TODO");
 	}
 	ASSERT(alignment > 0);
+	if (LLVMIsAGlobalVariable(llvm_value) && LLVMIsThreadLocal(llvm_value))
+	{
+		llvm_value = llvm_emit_call_intrinsic(c, intrinsic_id.threadlocal_address, &c->ptr_type, 1, &llvm_value, 1);
+	}
 	value->value = llvm_value;
 	value->alignment = alignment;
 	value->kind = BE_ADDRESS;
 	value->type = type_lowering(type);
 }
 
-void llvm_value_set_address_abi_aligned(BEValue *value, LLVMValueRef llvm_value, Type *type)
+void llvm_value_set_address_abi_aligned(GenContext *c, BEValue *value, LLVMValueRef llvm_value, Type *type)
 {
-	llvm_value_set_address(value, llvm_value, type_lowering(type), type_abi_alignment(type));
+	llvm_value_set_address(c, value, llvm_value, type_lowering(type), type_abi_alignment(type));
 }
 
 void llvm_value_addr(GenContext *c, BEValue *value)
@@ -70,13 +74,13 @@ void llvm_value_addr(GenContext *c, BEValue *value)
 		LLVMValueRef ref = llvm_add_global_raw(c, ".taddr", LLVMTypeOf(val), 0);
 		llvm_set_private_declaration(ref);
 		LLVMSetInitializer(ref, val);
-		llvm_value_set_address_abi_aligned(value, ref, value->type);
+		llvm_value_set_address_abi_aligned(c, value, ref, value->type);
 	}
 	else
 	{
 		LLVMValueRef temp = llvm_emit_alloca_aligned(c, value->type, "taddr");
 		llvm_store_to_ptr(c, temp, value);
-		llvm_value_set_address_abi_aligned(value, temp, value->type);
+		llvm_value_set_address_abi_aligned(c, value, temp, value->type);
 	}
 }
 
@@ -169,7 +173,7 @@ void llvm_value_set_decl_address(GenContext *c, BEValue *value, Decl *decl)
 {
 	ASSERT(!decl->is_value);
 	LLVMValueRef backend_ref = llvm_get_ref(c, decl);
-	llvm_value_set_address(value, backend_ref, decl->type, decl->alignment);
+	llvm_value_set_address(c, value, backend_ref, decl->type, decl->alignment);
 
 	if ((value->optional = llvm_get_opt_ref(c, decl)))
 	{