+++ releasenotes.md
@@ -9,6 +9,11 @@
 - `$typeof` may return a compile time type.
 - Improved error messages on missing qualifier on enum value. #2260
 - Add `--echo-prefix` to edit the prefix with `$echo` statements. Supports {FILE} and {LINE}
+- Catch accidental `foo == BAR;` where `foo = BAR;` was most likely intended. #2274
+- Improve error message when doing a rethrow in a function that doesn't return an optional.
+- Add `--list-asm` to view all supported `asm` instructions.
+- Formatting option "%h" now supports pointers.
+- Improve error on unsigned implicit conversion to signed.
 
 ### Fixes
 - mkdir/rmdir would not work properly with substring paths on non-windows platforms.
@@ -25,8 +30,18 @@
 - Const slice lengths were not always detected as constant.
 - Const slice indexing was not bounds checked.
 - Initialize pool correctly in print_backtrace.
+- `--max-mem` now works correctly again.
+- Casting a fault to a pointer would trigger an assert.
+- Make `to_float` more tolerant to spaces.
+- Fixes to thread local pointer handling.
+- Fixes to JSON parsing and Object.
+- Array indices are now using int64 internally.
+- Bit shift operation fails with inline uint enum despite matching underlying type #2279.
+- Fix to codegen when using a bitstruct constant defined using a cast with an operator #2248.
+- Function pointers are now compile time constants. 
 
 ### Stdlib changes
+- Improve contract for readline. #2280
 
 ## 0.7.3 Change list